refactor(e2e): port fixtures/helpers to ESM JS, wire playwright-bdd
Convert e2e/ to ESM with playwright-bdd v8. Adds support/ (server,
api, seed, fixtures) replacing fixtures/rdrs.ts + helpers/*.ts. Adds
currentUser fixture for per-scenario nanoid usernames. Old tests/
specs still present and unchanged in this commit; they fail-to-load
since their TS imports are gone — Task 2 deletes auth.spec.ts and
adds the first feature.
