+++ e2e/.gitignore
@@ -3,3 +3,4 @@ dist/
 playwright-report/
 test-results/
 blob-report/
+.features-gen/
