+++ e2e/support/api.js
@@ -1,12 +1,9 @@
-/** REST API helper for E2E tests. */
 export class ApiHelper {
-  constructor(private baseUrl: string) {}
+  constructor(baseUrl) {
+    this.baseUrl = baseUrl;
+  }
 
-  /** Register a new user. Idempotent: ignores 409 (already exists). */
-  async register(
-    username: string,
-    password: string
-  ): Promise<{ cookie: string }> {
+  async register(username, password) {
     const res = await fetch(`${this.baseUrl}/api/register`, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
@@ -19,11 +16,7 @@ export class ApiHelper {
     return { cookie: "" };
   }
 
-  /** Login and return the session cookie string. */
-  async login(
-    username: string,
-    password: string
-  ): Promise<{ cookie: string }> {
+  async login(username, password) {
     const res = await fetch(`${this.baseUrl}/api/session`, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
@@ -35,10 +28,7 @@ export class ApiHelper {
       throw new Error(`Login failed (${res.status}): ${body}`);
     }
     const setCookie = res.headers.getSetCookie?.() ?? [];
-    const cookie = setCookie
-      .map((c) => c.split(";")[0])
-      .join("; ");
+    const cookie = setCookie.map((c) => c.split(";")[0]).join("; ");
     return { cookie };
   }
-
 }
